# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Training Management System",
        "version": "1.0.0",
        "title": "APIs using Express Server",
        "contact": {
            "email": "vishal@successive.tech"
        }
    },
    "host": "localhost:9000",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/trainee": {
            "post": {
                "tags": [
                    "Trainee"
                ],
                "summary": "Add a new trainee",
                "description": "",
                "operationId": "addTrainee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Request body",
                        "description": "trainee details for creation",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TraineePost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/TraineePostRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Access",
                        "schema": {
                            "$ref": "#/definitions/TokenExpire"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Trainee"
                ],
                "summary": "Update an existing trainee",
                "description": "Updates an exisitng trainee",
                "operationId": "UpdateTrainee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Request body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TraineePut"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "$ref": "#/definitions/TraineePutRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized Access",
                        "schema": {
                            "$ref": "#/definitions/TokenExpire"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "get": {
                "tags": [
                    "Trainee"
                ],
                "summary": "Fetch all trainees",
                "description": "",
                "operationId": "findTrainee",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "description": "skip for pagination",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit for pagination",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An array of trainees",
                        "schema": {
                            "$ref": "#/definitions/TraineeGetRes"
                        }
                    },
                    "401": {
                        "description": "Token Expiration",
                        "schema": {
                            "$ref": "#/definitions/TokenExpire"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
...
        "TraineePutRes": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/TraineePut"
                },
                "message": {
                    "type": "string",
                    "example": "trainee updated succesfully"
                },
                "status": {
                    "type": "string",
                    "example": 200
                }
            }
        },
        "TraineePostRes": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/TraineePost"
                },
                "message": {
                    "type": "string",
                    "example": "trainee created succesfully"
                },
                "status": {
                    "type": "string",
                    "example": 200
                }
            }
        },
...
        "TraineeGetRes": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "string",
                    "example": "1"
                },
                "status": {
                    "type": "string",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "all trainees fetched"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/TraineeDetails"
                    }
                }
            }
        },
...
        "TraineeDetails": {
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "5d6f93dae4a9e958c46cf750"
                },
                "email": {
                    "type": "string",
                    "example": "vishal12@gmail.com"
                },
                "name": {
                    "type": "string",
                    "example": "divya"
                },
                "craetedBy": {
                    "type": "string",
                    "example": "5d6f4747b1b595320546b453"
                },
                "originalId": {
                    "type": "string",
                    "example": "5d6f93dae4a9e958c46cf750"
                },
                "role": {
                    "type": "string",
                    "example": "trainee"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2019-09-04T10:37:14.807Z"
                }
            }
        },
...
        "TraineePost": {
            "type": "object",
            "required": [
                "name",
                "email",
                "password"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Yuvraj"
                },
                "email": {
                    "type": "string",
                    "example": "yuvraj@gmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "123"
                }
            }
        },
        "TraineePut": {
            "type": "object",
            "required": [
                "id",
                "dataToUpdate"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "12334553Q"
                },
                "dataToUpdate": {
                    "type": "object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/TraineePost"
                        }
                    ]
                }
            }
        },
        "TokenExpire": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string",
                            "example": "Unauthorized access"
                        },
                        "status": {
                            "type": "string",
                            "example": 401
                        }
                    }
                },
                "timestamp": {
                    "type": "string",
                    "example": "2019-09-05T08:01:11.854Z"
                }
            }
        }
    }
}